+++ arrow-data/src/test/kotlin/arrow/data/UtilTest.kt
@@ -1,26 +1,8 @@
-/*
- * Copyright 2013 - 2016 Mario Arias
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package arrow.data
 
-package org.funktionale.utils
-
+import arrow.*
 import io.kotlintest.KTestJUnitRunner
 import io.kotlintest.matchers.shouldBe
-import arrow.UnitSpec
-import arrow.constant
-import arrow.identity
 import org.junit.runner.RunWith
 
 @RunWith(KTestJUnitRunner::class)
