--- funktionale-utils/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=Utils Module
-POM_ARTIFACT_ID=funktionale-utils
-POM_PACKAGING=jar
