--- funktionale-utils/src/main/kotlin/org/funktionale/utils/GetterOperation.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright 2013 - 2016 Mario Arias
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.funktionale.utils
-
-interface GetterOperation<in K, out V> {
-    val getter: (K) -> V
-    operator fun get(key: K): V = getter(key)
-}
-
-class GetterOperationImpl<in K, out V>(override val getter: (K) -> V) : GetterOperation<K, V>
-
